A bunch of work on keeping scopes up to date.
This is some hairy complex biz, but if I find myself doing this sort of thing
more than once, I'll factor out the core, which keeps a bunch of nested regions
up to date through arbitrary buffer edits. That's the hard part.

I still need to process newly inserted text to assign scopes to it, but the
bounds of the existing scopes are properly adjusted when insertions or
deletions happen.