An aborted attempt at rescoping inserted regions.
I think I'm going to toss this whole approach and revamp things completely,
because this is turning out to be fragile and wonky.

Having read through the TextMate source (insane amounts of incidental
complexity, yay C++), and the first-mate source (svelte and readable), I think
I'm going to tear apart the matcher tree so that I can record the matcher state
at the end of every line (like first-mate does it).

Then I can just record scopes on a line-relative basis, shift things down when
multiple lines are inserted or deleted, and re-process individual lines when
they are edited in any way.