Revamped grammar matching to be line based.
We now match a line at a time, and keep track of the matcher state at the end
of each line (ala first-mate). When the buffer is edited, we just completely
rematch the affected lines, starting from the first edited line and work our
way down.

During normal editing, this means we rematch the same line over and over again
as characters are typed. Matching is pretty efficient, so this is good enough
for now and probably forever.

I'm seeing some weirdness with the Java grammar, but that may be regexp
related, or something else, so I'll pinch this off here and investigate that
next.