Revamped to use -1 to indicate no match.
Some matches do not advance the match position, they just push new matchers
onto the stack, so we have to allow for a match where end == start (which we
previously took to indicate no match).

This means that it is now possible to infinite loop if a non-stack-pushing
matcher matches and advances zero, but we catch that at the source and ignore
such degenerate matches.