Hrm, we don't want scala-api to be provided.
A bunch of other packages depend on getting scala-api &c depends from
textmate-grammar.

I was originally excluding scala-api here because I didn't want all of that
showing up in the dependency graph inside Scaled when it loaded a package that
depended on textmate-grammar, but I already know how to solve that "properly"
and relying on Maven stuffs would never have worked anyway.