Preparations for custom grammar format.
Editing grammars embedded in plist XML is about as pleasant as stabbing myself
in the eye with a rusty sewing needle. I'm going to support a simple nested
dictionaries format (NDF) which uses indentation for structure and requires no
escaping (excepting trailing backslash for multiline values).

The grammars now print themselves in this format, so one will be able to load a
grammar from XML plist format and print it back out into a sane format for
subsequent editing.