Revamp rethinking to happen after fn dispatch.
Instead of doing a bunch of potentially redundant rethinking each time an edit
comes in, just note the potentially dirty lines during edits and when the fn
completes, process them all at once.

For simple edits this is basically equivalent, but when a fn results in
multiple changes to the buffer, this ensures that we only rethink after it's
all done, similarly if said fn is undone.

Presently this is suboptimal with regard to search/replace because it will
rethink all lines from the first replace to the last, whereas previously it
would rethink each line (and probably the one after) individually. Eventually
we're going to defer rethink of off-screen lines, so that won't end up being a
performance hit.